# tr-cuentas/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Transactions API
  description: API for managing accounts and movements
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /accounts:
    post:
      summary: Create a new account
      operationId: createAccount
      tags:
        - accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCuentaResponse'
    get:
      summary: Get all accounts
      operationId: getAllAccounts
      tags:
        - accounts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListCuentaResponse'
  /accounts/{id}:
    get:
      summary: Get account by ID
      operationId: getAccountById
      tags:
        - accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCuentaResponse'
    put:
      summary: Update account
      operationId: updateAccount
      tags:
        - accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCuentaResponse'
    delete:
      summary: Delete account
      operationId: deleteAccount
      tags:
        - accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /movements:
    post:
      summary: Create a new movement
      operationId: createMovement
      tags:
        - movements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoRequest'
      responses:
        '201':
          description: Movement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMovimientoResponse'
    get:
      summary: Get all movements
      operationId: getAllMovements
      tags:
        - movements
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListMovimientoResponse'
  /movements/{id}:
    get:
      summary: Get movement by ID
      operationId: getMovementById
      tags:
        - movements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMovimientoResponse'
    put:
      summary: Update movement
      operationId: updateMovement
      tags:
        - movements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMovimientoResponse'
    delete:
      summary: Delete movement
      operationId: deleteMovement
      tags:
        - movements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /reports/{clienteId}:
    get:
      summary: Get report for client
      operationId: getReport
      tags:
        - reports
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReporteEstadoCuentaResponse'
  /rclients:
    post:
      summary: Create or update client replica
      operationId: createOrUpdateReplica
      tags:
        - replica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /rclients/{id}:
    put:
      summary: Update client replica
      operationId: updateReplica
      tags:
        - replica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
    delete:
      summary: Delete client replica
      operationId: deleteReplica
      tags:
        - replica
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

components:
  schemas:
    CuentaRequest:
      type: object
      required:
        - numeroCuenta
        - tipoCuenta
        - saldoInicial
        - clienteId
      properties:
        numeroCuenta:
          type: string
          minLength: 6
        tipoCuenta:
          type: string
        saldoInicial:
          type: number
          minimum: 0
        estado:
          type: boolean
        clienteId:
          type: integer
          format: int64
    CuentaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numeroCuenta:
          type: string
        tipoCuenta:
          type: string
        saldoInicial:
          type: number
        estado:
          type: boolean
        clienteId:
          type: string
        nombreCliente:
          type: string
    MovimientoRequest:
      type: object
      required:
        - numeroCuenta
        - tipoMovimiento
        - valor
      properties:
        numeroCuenta:
          type: string
        tipoMovimiento:
          type: string
        valor:
          type: number
          minimum: 0
    MovimientoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        numeroCuenta:
          type: string
        tipoCuenta:
          type: string
        fecha:
          type: string
          format: date-time
        estado:
          type: boolean
        tipoMovimiento:
          type: string
        valor:
          type: number
        saldo:
          type: number
    ClienteRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        clienteId:
          type: string
        nombre:
          type: string
        estado:
          type: boolean
    ReporteEstadoCuentaResponse:
      type: object
      properties:
        clienteId:
          type: string
        nombreCliente:
          type: string
        rangoFechasSolicitado:
          type: string
        cuentas:
          type: array
          items:
            $ref: '#/components/schemas/ReporteCuenta'
    ReporteCuenta:
      type: object
      properties:
        numeroCuenta:
          type: string
        tipoCuenta:
          type: string
        saldoActual:
          type: number
        estado:
          type: boolean
        movimientos:
          type: array
          items:
            $ref: '#/components/schemas/ReporteMovimiento'
    ReporteMovimiento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fecha:
          type: string
          format: date-time
        tipoMovimiento:
          type: string
        valor:
          type: number
        saldo:
          type: number
    ApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
    ApiResponseCuentaResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CuentaResponse'
    ApiResponseListCuentaResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CuentaResponse'
    ApiResponseMovimientoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MovimientoResponse'
    ApiResponseListMovimientoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/MovimientoResponse'
    ApiResponseReporteEstadoCuentaResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ReporteEstadoCuentaResponse'
    ApiResponseBase:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
    ApiResponseError:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'